Add files via upload
I'm thrilled to share that I’ve recently begun learning Figma... In yesterday's session, we covered some exciting topics! First, we explored the 𝒕𝒂𝒊𝒍𝒘𝒊𝒏𝒅. 𝒄𝒐𝒏𝒇𝒊𝒈. 𝒕𝒔 file and learned how to configure the content array and customize the theme. The ability to define our own theme was a game-changer, and we saw how a simple example, like changing the button colour globally, can save time and maintain consistency. 🌟

Next, we tackled the issue with 𝒈𝒍𝒐𝒃𝒂𝒍. 𝒄𝒔𝒔 and figured out how to fix the errors we faced.

Then, we jumped into 𝑻𝒂𝒊𝒍𝒘𝒊𝒏𝒅 𝑪𝑺𝑺 itself and learned to customize layouts with background colours, flex, and more. We created a responsive header with a logo, navigation links, and cool hover effects using cursor-pointer and gap to perfect the spacing.

Finally, we engaged with 𝑭𝒊𝒈𝒎𝒂, learning how to design a website from scratch. Although some features have gone paid, our guide is finding a workaround for us!

We’ve been tasked with building a full website based on a Figma template. Time to put all these skills to the test!